refresh and history.goBack is no properly working, useEffect Warning is no ignored by eslint
