+++ frontend/src/context/SearchContextProvider.js
@@ -1,4 +1,4 @@
-import React, {useEffect, useState} from 'react';
+import React, {useState} from 'react';
 import SearchContext from "./SearchContext";
 import {
   getMovieSearchResultByName,
@@ -24,55 +24,6 @@ export default function SearchContextProvider({children}) {
   const voiceActorActors = [];
   const [actors,setActors] = useState([]);
 
-  useEffect(() => {
-      if (performance.navigation.type === performance.navigation.TYPE_RELOAD) {
-        console.info("This page is reloaded");
-        let currentPath = window.location.pathname.split("/");
-        console.log(window.location.pathname);
-        console.log(currentPath);
-        if (currentPath[1] !== "home") {
-          const previousSearch = currentPath[3].replace("%20", " ");
-          setSearchType(currentPath[2]);
-          switch (currentPath[2]) {
-
-            case "movie":
-              if (currentPath[1] === "search") {
-                console.log(currentPath[1]);
-                getMoviesByName(previousSearch);
-              } else {
-                console.log(currentPath[1]);
-                getMovieById(previousSearch);
-              }
-              break;
-
-            case "actor":
-              if (currentPath[1] === "search") {
-                getActorsByName(previousSearch);
-              } else {
-                getActorById(previousSearch);
-              }
-              break;
-
-            case "voiceactor":
-              if (currentPath[1] === "search") {
-                getVoiceActorByName(previousSearch);
-              } else {
-                getVoiceActorById(previousSearch);
-              }
-              break;
-
-            case "crew":
-              setName(previousSearch);
-              getMovieCrewByMovieId(currentPath[4]);
-              break;
-
-            default:
-              break;
-
-          }
-        }
-      }
-    },[])
 
   const getMoviesByName = (name) =>
     getMovieSearchResultByName(name).then((item) => setSearchItems(item));
