+++ frontend/src/searchPage/SearchPage.js
@@ -1,4 +1,4 @@
-import React, {useContext} from "react";
+import React, {useContext, useEffect} from "react";
 import { useHistory } from 'react-router-dom';
 import SearchList from "./SearchList";
 import SearchContext from "../context/SearchContext";
@@ -7,12 +7,13 @@ import DropdownButton from 'react-bootstrap/DropdownButton'
 
 export default function SearchPage() {
   const history = useHistory();
-  const {searchItems,name,setName,searchType,setSearchType,getActorsByName,getMoviesByName,setSearchItems,getVoiceActorByName,reloadStatus,setReloadStatus} = useContext(SearchContext);
+  const {name,setName,searchType,setSearchType,getActorsByName,getMoviesByName,setSearchItems,getVoiceActorByName,getMovieCrewByMovieId} = useContext(SearchContext);
 
-  if(searchItems.length === 0 && reloadStatus === true) {
-    setReloadStatus(false);
-    window.location.reload();
-  }
+  useEffect(() => {
+    onRefresh();
+    // eslint-disable-next-line
+  },[])
+
 
   if(searchType === "crew") {
     return (
@@ -37,13 +38,19 @@ export default function SearchPage() {
 
       <input name="name" disabled={!searchType} value={name} type="text" onChange={event => setName(event.target.value)}/>
       <button onClick={onSearch} disabled={!name}>Search</button>
-      <button onClick={onCancel}>Cancel</button>
+      <button onClick={goHome}>Cancel</button>
       <SearchList/>
     </>)
 
 
   function onCancel() {
     history.goBack();
+    setSearchItems([]);
+  }
+
+  function goHome() {
+    history.push("/");
+    setName("");
     setSearchItems([]);
   }
 
@@ -67,4 +74,38 @@ export default function SearchPage() {
         break;
     }
   }
+
+  function onRefresh() {
+
+    let currentPath = window.location.pathname.split("/");
+    if (currentPath.length > 2) {
+      const previousSearch = currentPath[3].replace("%20", " ");
+      setSearchType(currentPath[2]);
+      switch (currentPath[2]) {
+
+        case "movie":
+          getMoviesByName(previousSearch);
+          break;
+
+        case "actor":
+          getActorsByName(previousSearch);
+          break;
+
+        case "voiceactor":
+          getVoiceActorByName(previousSearch);
+          break;
+
+        case "crew":
+          setName(previousSearch);
+          getMovieCrewByMovieId(currentPath[4]);
+          break;
+
+        default:
+          break;
+
+      }
+    }
+  }
+
+
 }
